requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
